src/cloud_run/mimitalk_diagnosis_dev/requirements.txt
# Cloud Functionsの基本的なフレームワーク
functions-framework==3.*

# Vertex AI Agent Engineと通信するために必要
google-cloud-aiplatform

# Firestoreデータベースから会話ログを読み込むために必要
google-cloud-firestore